+++ src/app/rickAndMorty/components/rick-and-morty-characters-grid/grid-cell-renderer/character-history-cell/display-history-modal/display-history-modal.component.html
