+++ src/app/rickAndMorty/components/rick-and-morty-characters-grid/rick-and-morty-characters-grid.component.ts
@@ -8,8 +8,8 @@ import { takeUntilDestroyed, toSignal } from '@angular/core/rxjs-interop';
 import { AddNewCharacterButtonComponent } from '../add-new-character-button/add-new-character-button.component';
 import { CharacterHistoryCellComponent } from './grid-cell-renderer/character-history-cell/character-history-cell.component';
 import { UpdateCharacterCellComponent } from './grid-cell-renderer/update-character-cell/update-character-cell.component';
-import { UpdateRickAndMortyCharacterModalComponent } from '../update-rick-and-morty-character-modal/update-rick-and-morty-character-modal.component';
-import { DisplayHistoryModalComponent } from '../display-history-modal/display-history-modal.component';
+import { UpdateRickAndMortyCharacterModalComponent } from './grid-cell-renderer/update-character-cell/update-rick-and-morty-character-modal/update-rick-and-morty-character-modal.component';
+import { DisplayHistoryModalComponent } from './grid-cell-renderer/character-history-cell/display-history-modal/display-history-modal.component';
 
 @Component({
   selector: 'app-rick-and-morty-characters-grid',
